requirements.txt
requests
websocket-client
sortedcontainers
python-dateutil
# Optional Dependencies (used for examples and better logging for historical_data_downloader)
absl-py
progressbar2